moved proof of ownership request to WalletAuthorizedRequest from WalletUnauthorizedRequest
