+++ app/src/main/java/com/babylon/wallet/android/domain/model/messages/WalletUnauthorizedRequest.kt
@@ -1,8 +1,5 @@
 package com.babylon.wallet.android.domain.model.messages
 
-import com.radixdlt.sargon.AccountAddress
-import com.radixdlt.sargon.Exactly32Bytes
-import com.radixdlt.sargon.IdentityAddress
 import com.radixdlt.sargon.WalletInteractionId
 
 data class WalletUnauthorizedRequest(
@@ -10,15 +7,8 @@ data class WalletUnauthorizedRequest(
     override val interactionId: WalletInteractionId,
     val requestMetadata: RequestMetadata,
     val oneTimeAccountsRequestItem: AccountsRequestItem? = null,
-    val oneTimePersonaDataRequestItem: PersonaDataRequestItem? = null,
-    val proofOfOwnershipRequestItem: ProofOfOwnershipRequestItem? = null
+    val oneTimePersonaDataRequestItem: PersonaDataRequestItem? = null
 ) : DappToWalletInteraction(remoteEntityId, interactionId, requestMetadata) {
-
-    data class ProofOfOwnershipRequestItem(
-        val challenge: Exactly32Bytes,
-        val accountAddresses: List<AccountAddress>? = null,
-        val personaAddress: IdentityAddress? = null
-    )
 
     fun isValidRequest(): Boolean {
         return oneTimeAccountsRequestItem?.isValidRequestItem() != false
